language/assignment-3---rnns-for-text-generation-julifurjes-main/requirements.txt
matplotlib
numpy
pandas
tensorflow-cpu
joblib